Create C++ Project Car and Bike login system using C++, OOP concepts
This project has a login and signup system for the user, this  program first asks the 
user to log in or sign up, for signing up user cannot take a name that is already 
taken by another user, after signing up user can choose a vehicle which he/she 
want to select after that user can delete the booking and can re-book a vehicle 
again. Users can also check their detail of booking and ID. We have also made 
logout button which clears the screen. If a vehicle is already booked by another 
user, the program prints that “This vehicle is already Taken by someone, please 
select another vehicle”.
I haven't used file handling but file handling can be used to store details for the long term and databases can also be attached.